Move definition(s) of union ubuf to lmptype.h

+++ src/MANYBODY/pair_eam_cd.cpp
@@ -518,7 +518,7 @@ void PairEAMCD::read_h_coeff(char *filename)
     int degree = values.next_int();
     nhcoeff = degree+1;
 
-    if (values.count() != nhcoeff + 1 || nhcoeff < 1)
+    if ((int)values.count() != nhcoeff + 1 || nhcoeff < 1)
       error->one(FLERR, "Failed to read h(x) function coefficients in EAM file.");
 
     hcoeff = new double[nhcoeff];
